Fixed warning C4267: 'initializing': conversion from 'size_t' to 'int', possible loss of data

+++ src/dialog/windows/SDL_windowsdialog.c
@@ -132,7 +132,7 @@ void windows_ShowFileDialog(void *ptr)
         return;
     }
 
-    int filter_len = SDL_strlen(filterlist);
+    int filter_len = (int)SDL_strlen(filterlist);
 
     for (char *c = filterlist; *c; c++) {
         if (*c == '\x01') {
